Removed index.js from main src file. Renamed interpreter package to parser to reflect content. Renamed filesearch package to locator to reflect content. Restructured affected tests and artifacts.
