--- src/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-exports.filesearch = require('./filesearch');
-exports.interpreter = require('./interpreter');
-exports.model = require('./model');
