+++ src/locator/index.js
